Returns the mean and CI

Generates mean and confidence intervals using bootstrap.

Usage

mean_ci(
  window,
  conf = 0.95,
  na.rm = TRUE,
  type = "basic",
  R = 1000,
  parallel = "no",
  ncpus = getOption("boot.ncpus", 1L),
  cl = NULL
)

Arguments

window

vector of data values

conf

confidence level of the required interval. NA if skipping calculating the bootstrapped CI

na.rm

logical TRUE/FALSE. Remove NAs from the dataset. Defaults TRUE

type

character string, one of c("norm","basic", "stud", "perc", "bca"). "all" is not a valid value. See boot.ci

R

the number of bootstrap replicates. see boot

parallel

The type of parallel operation to be used (if any). see boot

ncpus

integer: number of process to be used in parallel operation. see boot

cl

optional parallel or snow cluster for use if parallel = "snow". see boot

Value

named list with mean and (optionally) specified confidence interval